+++ src/app/src/advancedsetupdialog.h
@@ -32,6 +32,7 @@ class AdvancedSetupDialog : public QDialog{
         void sig_ni_dev_change(std::string m_niDev, std::string m_trigDev);
         void sig_trigger_mode_change(int16_t triggerMode);
         void sig_enable_live_view_during_acquisition_change(bool enable);
+        void sig_display_rois_during_live_view_change(bool enable);
         void sig_close_adv_settings();
 
     private slots:
@@ -40,6 +41,7 @@ class AdvancedSetupDialog : public QDialog{
         void on_triggerDeviceList_currentTextChanged(const QString &text);
         void on_triggerModeList_currentTextChanged(const QString &text);
         void on_checkEnableLiveViewDuringAcq_stateChanged(int state);
+        void on_checkDisplayRoisDuringLiveView_stateChanged(int state);
         void on_checkDownsampleRawFiles_stateChanged(int state);
         void on_checkKeepOriginalRaw_stateChanged(int state);
         void on_binFactorList_currentTextChanged(const QString &text);
@@ -54,6 +56,7 @@ class AdvancedSetupDialog : public QDialog{
         std::string m_trigDev;
         int16_t m_triggerMode;
         bool m_enableLiveViewDuringAcquisition;
+        bool m_displayRoisDuringLiveView;
         bool m_enableDownsampleRawFiles;
         bool m_keepOriginalRaw;
         uint8_t m_binFactor;