+++ src/app/src/config.h
@@ -59,6 +59,7 @@ class Config {
 
         //acquisition.live_view options
         bool enableLiveViewDuringAcquisition;
+        bool displayRoisDuringLiveView;
         bool vflip;
         bool hflip;